requirements.txt
astroid==2.15.6
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.9
dill==0.3.7
distlib==0.3.7
django-upgrade==1.14.1
djlint==1.32.1
EditorConfig==0.12.3
filelock==3.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
isort==5.12.0
jsbeautifier==1.14.9
json5==0.9.14
lazy-object-proxy==1.9.0
mccabe==0.7.0
pathspec==0.11.2
platformdirs==3.10.0
pylint==2.17.5
PyYAML==6.0.1
regex==2023.8.8
six==1.16.0
tokenize-rt==5.2.0
tomlkit==0.12.1
tqdm==4.66.1
virtualenv==20.24.3
wrapt==1.15.0
